requirements.txt
amqp==2.6.0
anyjson==0.3.3
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
ecdsa==0.15
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
oauthlib==3.1.0
plaid-python==4.1.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
python-jose==3.1.0
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
